requirements.txt
Flask==2.0.2
tensorflow==2.6.0
Pillow==8.4.0
